Refactor. Split cli.ts
