+++ packages/compiler/src/core/cli/actions/format.ts
@@ -1,0 +1,28 @@
+import { findUnformattedTypeSpecFiles, formatTypeSpecFiles } from "../../formatter-fs.js";
+
+export interface FormatArgs {
+  include: string[];
+  exclude?: string[];
+  debug?: boolean;
+  check?: boolean;
+}
+export async function formatAction(args: FormatArgs) {
+  if (args["check"]) {
+    const unformatted = await findUnformattedTypeSpecFiles(args["include"], {
+      exclude: args["exclude"],
+      debug: args.debug,
+    });
+    if (unformatted.length > 0) {
+      console.log(`Found ${unformatted.length} unformatted files:`);
+      for (const file of unformatted) {
+        console.log(` - ${file}`);
+      }
+      process.exit(1);
+    }
+  } else {
+    await formatTypeSpecFiles(args["include"], {
+      exclude: args["exclude"],
+      debug: args.debug,
+    });
+  }
+}
