+++ packages/compiler/src/core/cli/index.ts
@@ -1,1 +1,1 @@
-export * from "./args.js";
+export * from "./actions/compile/args.js";
